Internal soldering of Force Trainer
Other devices have a similar Neurosky board. 

They *might* be setup to use a different baud
 rate (i.e., code might need to be modified).

Look for the "T" (next to the "R") terminal,
and solder to that to obtain a signal.

Verify that you have not shorted any of the
adjacent pads to the pad that you tack onto.

Close case, trapping pair of wires between battery
compartment, and the case, to form a strain relief
(i.e., to prevent pulling on the solder joint).